Refactor for execution model.
