+++ daft/logical/builder.py
@@ -21,7 +21,10 @@ from daft.runners.partitioning import PartitionCacheEntry
 if TYPE_CHECKING:
     from pyiceberg.table import Table as IcebergTable
 
-    from daft.plan_scheduler.physical_plan_scheduler import AdaptivePhysicalPlanScheduler, PhysicalPlanScheduler
+    from daft.plan_scheduler.physical_plan_scheduler import (
+        AdaptivePhysicalPlanScheduler,
+        PhysicalPlanScheduler,
+    )
 
 
 class LogicalPlanBuilder:
@@ -41,14 +44,22 @@ class LogicalPlanBuilder:
         """
         from daft.plan_scheduler.physical_plan_scheduler import PhysicalPlanScheduler
 
-        return PhysicalPlanScheduler(self._builder.to_physical_plan_scheduler(daft_execution_config))
+        return PhysicalPlanScheduler.from_logical_plan_builder(
+            self,
+            daft_execution_config,
+        )
 
     def to_adaptive_physical_plan_scheduler(
         self, daft_execution_config: PyDaftExecutionConfig
     ) -> AdaptivePhysicalPlanScheduler:
-        from daft.plan_scheduler.physical_plan_scheduler import AdaptivePhysicalPlanScheduler
+        from daft.plan_scheduler.physical_plan_scheduler import (
+            AdaptivePhysicalPlanScheduler,
+        )
 
-        return AdaptivePhysicalPlanScheduler(self._builder.to_adaptive_physical_plan_scheduler(daft_execution_config))
+        return AdaptivePhysicalPlanScheduler.from_logical_plan_builder(
+            self,
+            daft_execution_config,
+        )
 
     def schema(self) -> Schema:
         """
@@ -78,10 +89,20 @@ class LogicalPlanBuilder:
 
     @classmethod
     def from_in_memory_scan(
-        cls, partition: PartitionCacheEntry, schema: Schema, num_partitions: int, size_bytes: int, num_rows: int
+        cls,
+        partition: PartitionCacheEntry,
+        schema: Schema,
+        num_partitions: int,
+        size_bytes: int,
+        num_rows: int,
     ) -> LogicalPlanBuilder:
         builder = _LogicalPlanBuilder.in_memory_scan(
-            partition.key, partition, schema._schema, num_partitions, size_bytes, num_rows
+            partition.key,
+            partition,
+            schema._schema,
+            num_partitions,
+            size_bytes,
+            num_rows,
         )
         return cls(builder)
 
@@ -125,7 +146,11 @@ class LogicalPlanBuilder:
         return LogicalPlanBuilder(builder)
 
     def unpivot(
-        self, ids: list[Expression], values: list[Expression], variable_name: str, value_name: str
+        self,
+        ids: list[Expression],
+        values: list[Expression],
+        variable_name: str,
+        value_name: str,
     ) -> LogicalPlanBuilder:
         ids_pyexprs = [expr._expr for expr in ids]
         values_pyexprs = [expr._expr for expr in values]
