+++ src/daft-plan/src/lib.rs
@@ -9,7 +9,7 @@ mod logical_ops;
 mod logical_optimization;
 mod logical_plan;
 mod partitioning;
-mod physical_ops;
+pub mod physical_ops;
 mod physical_plan;
 mod physical_planner;
 mod resource_request;
@@ -24,12 +24,18 @@ pub use daft_core::join::{JoinStrategy, JoinType};
 use daft_scan::file_format::FileFormat;
 pub use logical_plan::{LogicalPlan, LogicalPlanRef};
 pub use partitioning::ClusteringSpec;
-pub use physical_plan::PhysicalPlanScheduler;
+pub use physical_plan::{PhysicalPlan, PhysicalPlanRef};
+pub use physical_planner::{
+    logical_to_physical, AdaptivePlanner, MaterializedResults, QueryStageOutput,
+};
 pub use resource_request::ResourceRequest;
-pub use source_info::{FileInfo, FileInfos};
+pub use sink_info::OutputFileInfo;
+pub use source_info::{FileInfo, FileInfos, InMemoryInfo};
 
 #[cfg(feature = "python")]
 use pyo3::prelude::*;
+#[cfg(feature = "python")]
+pub use sink_info::{DeltaLakeCatalogInfo, IcebergCatalogInfo};
 #[cfg(feature = "python")]
 use {
     daft_scan::file_format::{
@@ -42,8 +48,6 @@ use {
 pub fn register_modules(_py: Python, parent: &PyModule) -> PyResult<()> {
     use daft_scan::file_format::DatabaseSourceConfig;
 
-    use crate::physical_planner::python::AdaptivePhysicalPlanScheduler;
-
     parent.add_class::<PyLogicalPlanBuilder>()?;
     parent.add_class::<FileFormat>()?;
     parent.add_class::<PyFileFormatConfig>()?;
@@ -51,8 +55,6 @@ pub fn register_modules(_py: Python, parent: &PyModule) -> PyResult<()> {
     parent.add_class::<JsonSourceConfig>()?;
     parent.add_class::<CsvSourceConfig>()?;
     parent.add_class::<DatabaseSourceConfig>()?;
-    parent.add_class::<PhysicalPlanScheduler>()?;
-    parent.add_class::<AdaptivePhysicalPlanScheduler>()?;
     parent.add_class::<ResourceRequest>()?;
     parent.add_class::<FileInfos>()?;
     parent.add_class::<FileInfo>()?;
