+++ src/daft-plan/src/physical_ops/explode.rs
@@ -81,7 +81,7 @@ mod tests {
 
     use crate::{
         partitioning::{HashClusteringConfig, UnknownClusteringConfig},
-        physical_planner::plan,
+        physical_planner::logical_to_physical,
         test::{dummy_scan_node, dummy_scan_operator},
         ClusteringSpec,
     };
@@ -100,7 +100,7 @@ mod tests {
         .explode(vec![col("b")])?
         .build();
 
-        let physical_plan = plan(logical_plan, cfg)?;
+        let physical_plan = logical_to_physical(logical_plan, cfg)?;
 
         let expected_clustering_spec =
             ClusteringSpec::Hash(HashClusteringConfig::new(3, vec![col("a")]));
@@ -127,7 +127,7 @@ mod tests {
         .explode(vec![col("b")])?
         .build();
 
-        let physical_plan = plan(logical_plan, cfg)?;
+        let physical_plan = logical_to_physical(logical_plan, cfg)?;
 
         let expected_clustering_spec = ClusteringSpec::Unknown(UnknownClusteringConfig::new(3));
