+++ src/daft-plan/src/physical_planner/planner.rs
@@ -3,6 +3,7 @@ use std::sync::Arc;
 use common_daft_config::DaftExecutionConfig;
 use common_error::DaftResult;
 use common_treenode::{Transformed, TreeNode, TreeNodeRewriter, TreeNodeVisitor};
+use serde::{Deserialize, Serialize};
 
 use crate::logical_ops::Source;
 use crate::logical_optimization::Optimizer;
@@ -243,7 +244,8 @@ impl TreeNodeRewriter for ReplacePlaceholdersWithMaterializedResult {
     }
 }
 
-pub(super) enum QueryStageOutput {
+#[derive(Debug, Serialize, Deserialize)]
+pub enum QueryStageOutput {
     Partial {
         physical_plan: PhysicalPlanRef,
         source_id: usize,
@@ -263,6 +265,13 @@ impl QueryStageOutput {
             QueryStageOutput::Final { physical_plan } => (None, physical_plan),
         }
     }
+
+    pub fn source_id(&self) -> Option<usize> {
+        match self {
+            QueryStageOutput::Partial { source_id, .. } => Some(*source_id),
+            QueryStageOutput::Final { .. } => None,
+        }
+    }
 }
 #[derive(PartialEq, Debug)]
 enum AdaptivePlannerStatus {
@@ -271,12 +280,12 @@ enum AdaptivePlannerStatus {
     Done,
 }
 
-pub(super) struct MaterializedResults {
+pub struct MaterializedResults {
     pub source_id: usize,
     pub in_memory_info: InMemoryInfo,
 }
 
-pub(super) struct AdaptivePlanner {
+pub struct AdaptivePlanner {
     logical_plan: LogicalPlanRef,
     cfg: Arc<DaftExecutionConfig>,
     status: AdaptivePlannerStatus,
@@ -291,7 +300,7 @@ impl AdaptivePlanner {
         }
     }
 
-    pub fn next(&mut self) -> DaftResult<QueryStageOutput> {
+    pub fn next_stage(&mut self) -> DaftResult<QueryStageOutput> {
         assert_eq!(self.status, AdaptivePlannerStatus::Ready);
 
         let mut rewriter = QueryStagePhysicalPlanTranslator {
