+++ src/daft-scheduler/src/adaptive.rs
@@ -3,11 +3,15 @@ use std::sync::Arc;
 use common_daft_config::DaftExecutionConfig;
 use daft_core::schema::Schema;
 
-use crate::physical_planner::planner::MaterializedResults;
-use crate::source_info::InMemoryInfo;
-use crate::LogicalPlan;
-use crate::{physical_planner::planner::AdaptivePlanner, PhysicalPlanScheduler};
-use pyo3::prelude::*;
+use crate::PhysicalPlanScheduler;
+use daft_plan::InMemoryInfo;
+use daft_plan::LogicalPlan;
+use daft_plan::{AdaptivePlanner, MaterializedResults};
+
+#[cfg(feature = "python")]
+use {
+    common_daft_config::PyDaftExecutionConfig, daft_plan::PyLogicalPlanBuilder, pyo3::prelude::*,
+};
 /// A work scheduler for physical plans.
 #[cfg_attr(feature = "python", pyclass(module = "daft.daft"))]
 pub struct AdaptivePhysicalPlanScheduler {
@@ -25,11 +29,25 @@ impl AdaptivePhysicalPlanScheduler {
 #[cfg(feature = "python")]
 #[pymethods]
 impl AdaptivePhysicalPlanScheduler {
+    #[staticmethod]
+    pub fn from_logical_plan_builder(
+        logical_plan_builder: &PyLogicalPlanBuilder,
+        py: Python<'_>,
+        cfg: PyDaftExecutionConfig,
+    ) -> PyResult<Self> {
+        py.allow_threads(|| {
+            let logical_plan = logical_plan_builder.builder.build();
+            Ok(AdaptivePhysicalPlanScheduler::new(
+                logical_plan,
+                cfg.config.clone(),
+            ))
+        })
+    }
     pub fn next(&mut self, py: Python) -> PyResult<(Option<usize>, PhysicalPlanScheduler)> {
         py.allow_threads(|| {
-            let output = self.planner.next()?;
-            let (sid, pps) = output.unwrap();
-            Ok((sid, pps.into()))
+            let output = self.planner.next_stage()?;
+            let sid = output.source_id();
+            Ok((sid, output.into()))
         })
     }
